update project to target .NET Standard and add nuke build  for build system
