+++ .travis.yml
@@ -1,0 +1,8 @@
+dist: xenial
+sudo: required
+language: csharp
+mono: none
+dotnet: 2.2.101
+solution: NepaliDateConverter.sln
+script:
+  - dotnet run  --project ./build/_build.csproj
