--- NepaliDateConverter/NepaliDateConverter.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageId>NepaliDateConverter.NETCORE</PackageId>
-  </PropertyGroup>
-
-</Project>
