+++ src/NepaliDateConverter/NepaliDateConverter.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard1.0</TargetFramework>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageId>NepaliDateConverter.NETCORE</PackageId>
+  </PropertyGroup>
+
+</Project>
